\name{demo}
\alias{demo}
\title{program check}
\usage{
demo(foldername)
}
\arguments{
\item{in_foldername}{folder which cotain barcodes.tsv, features.tsv, matirx.mtx}
\item{out_foldername}{output path}
\item{fragment}{fragment file}
\item{chr}{chromatin which}
in_foldername,out_foldername,fragment,chr,from,to,core,width=6,height=4
}
\description{
clustering cell
}

