\name{demo}
\alias{demo}
\title{program check}
\usage{
demo(foldername)
}
in_foldername,out_foldername,fragment,chr,from,to,core,width=6,height=4,
\arguments{
\item{in_foldername}{folder which cotain barcodes.tsv, features.tsv, matirx.mtx}
\item{out_foldername}{output path}
\item{fragment}{fragment file}
\item{macs2path}{macs2path}
\item{bedtoolspath}{bedtoolspath}
\item{chr}{chromatin which}
\item{from,to}{}
\item{core}{The number of core to use}
\item{width,height}{output figure size}
}
\description{
clustering cell and detecting reguretory relationship
}

