\name{demo}
\alias{demo}
\title{program check}
\usage{
demo(in_foldername,out_foldername,fragment,macs2path,bedtoolspath,awkpath,core)
}
\arguments{
\item{in_foldername}{folder which cotain barcodes.tsv, features.tsv, matirx.mtx}
\item{out_foldername}{output path}
\item{fragment}{fragment file}
\item{macs2path}{macs2path}
\item{bedtoolspath}{bedtoolspath}
\item{core}{The number of core to use}
}
\description{
clustering cell and detecting reguretory relationship
}

